refactor: 이동 전략 패턴을 동작 파라미터화로 리팩토링
