+++ src/main/java/step3/domain/strategy/MovementStrategy.java
@@ -1,0 +1,6 @@
+package step3.domain.strategy;
+
+@FunctionalInterface
+public interface MovementStrategy {
+    boolean isMovable();
+}
