--- src/main/java/step3/domain/strategy/movement/MovementStrategy.java
@@ -1,6 +1,0 @@
-package step3.domain.strategy.movement;
-
-@FunctionalInterface
-public interface MovementStrategy {
-    boolean isMovable(final int movementStrategyNumber);
-}
